Truncate long list items
For long menu items, truncate the item with an ellipsis.
This is the implementation currently in Next.ft.com, but should probably be
revisited at some point as it would be better to wrap long items. The problem
with wrapping is it causes funny behaviour on lists that split in two
and float next to one another.